backend/requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Smorest==0.42.0


# Database
SQLAlchemy==2.0.23
psycopg==3.1.18
alembic==1.13.1

# Validation and serialization
marshmallow==3.20.1
pydantic==2.5.0
jinja2==3.1.2

# External services
stripe==7.8.0
redis==5.0.1
celery==5.3.4
sendgrid==6.10.0
twilio==8.10.0

# Authentication and security
PyJWT==2.8.0
cryptography>=42.0.5

# HTTP client
httpx==0.25.2
requests==2.31.0

# Background tasks
tenacity==8.2.3

# Timezone handling
pytz==2023.3

# Logging and monitoring
structlog==23.2.0
sentry-sdk[flask]==1.38.0
prometheus-client==0.17.1
psutil==5.9.6

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
python-dotenv==1.0.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Production
gunicorn==21.2.0
